Move away from pect-extension
This commit moves away from using the PECL-extension towards using a
self-created GnuPG interface that itself agains mimics the functions of
the GnuPG extension.

This should make testing and execution without GnuPG being available in
either an extension or as binary easier.
