+++ src/GnuPG/ErrorHandler.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace PharIo\GnuPG\GnuPG;
+
+interface ErrorHandler
+{
+    /**
+     * @return false|string
+     */
+    public function geterror();
+}
