+++ src/GnuPG/Verifier.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace PharIo\GnuPG\GnuPG;
+
+interface Verifier
+{
+    public function verify(string $content, string $signature): bool;
+}
